L2Math  * increasing precision and preventing overflow at distance calculation
git-svn-id: http://svn.l2jfree.com/l2jfree@5663 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/l2j-commons/src/main/java/com/l2jfree/lang/L2Math.java
@@ -85,68 +85,68 @@ public final class L2Math
 	
 	public static double calculateDistance(int x1, int y1, int x2, int y2)
 	{
-		final int diffX = x1 - x2;
-		final int diffY = y1 - y2;
+		final long diffX = x1 - x2;
+		final long diffY = y1 - y2;
 		
 		return Math.sqrt(diffX * diffX + diffY * diffY);
 	}
 	
 	public static double calculateDistance(int x1, int y1, int z1, int x2, int y2, int z2)
 	{
-		final int diffX = x1 - x2;
-		final int diffY = y1 - y2;
-		final int diffZ = z1 - z2;
+		final long diffX = x1 - x2;
+		final long diffY = y1 - y2;
+		final long diffZ = z1 - z2;
 		
 		return Math.sqrt(diffX * diffX + diffY * diffY + diffZ * diffZ);
 	}
 	
-	public static int calculateDistanceSq(int x1, int y1, int x2, int y2)
+	public static long calculateDistanceSq(int x1, int y1, int x2, int y2)
 	{
-		final int diffX = x1 - x2;
-		final int diffY = y1 - y2;
+		final long diffX = x1 - x2;
+		final long diffY = y1 - y2;
 		
 		return diffX * diffX + diffY * diffY;
 	}
 	
-	public static int calculateDistanceSq(int x1, int y1, int z1, int x2, int y2, int z2)
+	public static long calculateDistanceSq(int x1, int y1, int z1, int x2, int y2, int z2)
 	{
-		final int diffX = x1 - x2;
-		final int diffY = y1 - y2;
-		final int diffZ = z1 - z2;
+		final long diffX = x1 - x2;
+		final long diffY = y1 - y2;
+		final long diffZ = z1 - z2;
 		
 		return diffX * diffX + diffY * diffY + diffZ * diffZ;
 	}
 	
 	public static boolean isDistanceLessThan(int x1, int y1, int x2, int y2, int limit)
 	{
-		final int diffX = x1 - x2;
-		final int diffY = y1 - y2;
+		final long diffX = x1 - x2;
+		final long diffY = y1 - y2;
 		
 		return diffX * diffX + diffY * diffY <= limit * limit;
 	}
 	
 	public static boolean isDistanceLessThan(int x1, int y1, int z1, int x2, int y2, int z2, int limit)
 	{
-		final int diffX = x1 - x2;
-		final int diffY = y1 - y2;
-		final int diffZ = z1 - z2;
+		final long diffX = x1 - x2;
+		final long diffY = y1 - y2;
+		final long diffZ = z1 - z2;
 		
 		return diffX * diffX + diffY * diffY + diffZ * diffZ <= limit * limit;
 	}
 	
-	public static boolean isDistanceLessThanSq(int x1, int y1, int x2, int y2, int limitSq)
+	public static boolean isDistanceLessThanSq(int x1, int y1, int x2, int y2, long limitSq)
 	{
-		final int diffX = x1 - x2;
-		final int diffY = y1 - y2;
+		final long diffX = x1 - x2;
+		final long diffY = y1 - y2;
 		
 		return diffX * diffX + diffY * diffY <= limitSq;
 	}
 	
-	public static boolean isDistanceLessThanSq(int x1, int y1, int z1, int x2, int y2, int z2, int limitSq)
+	public static boolean isDistanceLessThanSq(int x1, int y1, int z1, int x2, int y2, int z2, long limitSq)
 	{
-		final int diffX = x1 - x2;
-		final int diffY = y1 - y2;
-		final int diffZ = z1 - z2;
+		final long diffX = x1 - x2;
+		final long diffY = y1 - y2;
+		final long diffZ = z1 - z2;
 		
 		return diffX * diffX + diffY * diffY + diffZ * diffZ <= limitSq;
 	}
@@ -157,12 +157,12 @@ public final class L2Math
 	 * @return <code>base<sup>exponent</sup></code>
 	 * @throws IllegalArgumentException if the exponent is negative
 	 */
-	public static int pow(final int base, int exponent)
+	public static long pow(final int base, int exponent)
 	{
 		if (exponent < 0)
 			throw new IllegalArgumentException("Exponent must be non-negative!");
 		
-		int result = 1;
+		long result = 1;
 		
 		while (exponent-- > 0)
 			result *= base;
